gateway/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
httpx
python-decouple
